pyproject.toml
[tool.poetry]
name = "fastapi-profiling-example"
version = "1.0.0"
description = "fastapi-profiling-example"
authors = ["Thibault Ayanides <thibault.serti@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
pyroscope-io = "^0.8.5"

[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.8"
black = "^23.1.0"
pylint = "^2.17.0"
mypy = "^1.1.1"
isort = "^5.12.0"
pytest = "^7.2.2"
coverage = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"
force_single_line = false
atomic = false
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
filter_files = true
known_first_party = "poetry"

[tool.mypy]
ignore_missing_imports = true


[tool.pylint.messages_control]
max-line-length = 120

extension-pkg-whitelist = "pydantic"

disable = [
]
